requirements.txt
aiohttp==3.6.2
antiorm==1.2.1
APScheduler==3.6.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
config==0.5.0.post0
db==0.1.1
discord.py==1.3.4
idna==2.9
multidict==4.7.6
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tzlocal==2.1
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
